refactor(provider/kubernetes): Move v1 cache code into v1 package (#1862)
