+++ clouddriver-kubernetes/src/main/groovy/com/netflix/spinnaker/clouddriver/kubernetes/v1/provider/view/KubernetesClusterProvider.groovy
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.clouddriver.kubernetes.provider.view
+package com.netflix.spinnaker.clouddriver.kubernetes.v1.provider.view
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.netflix.frigga.Names
